Inverted P and Q for KLD computation, more fixes
- P is now the program's distribution and Q is the RNN's

A set of paths with calls, branches and locations is first extracted
from the program. For each location l, we get the set of paths that
terminate at l: {p1, p2, ... pk}. For each pi,

Pr(pi) = 1/b1 * 1/b2 * ... 1/bm where bj are the branches in pi

To make this a distribution over the set {p1,..pk}, Pr(pi) is
normalized.

Now consider the domain of all possible sequences of calls in the
universe. There are two distributions on this domain -- P that is
imposed by the paths {p1,..pk} and Q that is imposed by the RNN. Note
that this domain is only on call sequences, and not locations or
branches.

For each sequence s = [c1 c2 ... cn] (calls) in this domain, we have

P(s) = Sum_k Pr(s|pk) * Pr(pk)

where Pr(s|pi) = 0 if pi does not produce the call sequence s, or 1
otherwise. Pr(pi) is as defined above.

Q(s) = Pr(c1) * Pr(c2) * ... Pr(cn)

where Pr(ci) is given by the RNN at each step i.

Now that P and Q are defined on the same domain of call sequences, we
have:

KLD(l) = Sum_s P(s) * log( P(s)/Q(s) )